Implemented continuously running tasks that: starting from the start date repeat forever through a given interval.
These get executes through scheduler process and are very similar to cron processes, but manageable from PHP.

+++ src/ShiftGearman/GearmanService.php
@@ -364,7 +364,7 @@ class GearmanService
         {
             $task->markRepeatedOnce();
 
-            if(1 > $task->getRepeatTimes())
+            if(1 > $task->getRepeatTimes() && !$task->isContinuous())
                 $this->getSchedulerRepository()->delete($task, false);
             else
                 $this->getSchedulerRepository()->save($task, false);
